+++ operators/ARMORED_control_loops.py
@@ -1,10 +1,12 @@
+version = (1, 1, 0)
+
 import bpy
 
 
 class MESH_OT_armored_control_loops(bpy.types.Operator):
 	'''Use the bevel tool to create control loops for subdivision surfaces.
 
-armoredColony.com '''
+	armoredColony.com '''
 
 	bl_idname = 'mesh.armored_control_loops'
 	bl_label = 'ARMORED Control Loops'
@@ -15,7 +17,7 @@ armoredColony.com '''
 		return context.active_object is not None
 
 	def execute(self, context):
-		bpy.ops.mesh.armored_select_hard_edges()	# EXTERNAL SCRIPT.
+		# bpy.ops.mesh.armored_select_hard_edges()	# EXTERNAL SCRIPT.
 		bpy.ops.mesh.bevel('INVOKE_DEFAULT', segments=2, profile=1, miter_outer='ARC')
 
 		return {'FINISHED'}